+++ test/pokemon_store/stubs/Logging.h
@@ -3,3 +3,7 @@
 #define LOG_ERR(...) \
   do {               \
   } while (false)
+
+#define LOG_INF(...) \
+  do {               \
+  } while (false)